Move classes and split DispatcherServlet config
