+++ src/main/java/de/vorb/platon/web/api/ApiConfig.java
@@ -16,16 +16,24 @@
 
 package de.vorb.platon.web.api;
 
+import de.vorb.platon.persistence.PersistenceConfig;
+import de.vorb.platon.security.SecurityConfig;
+import de.vorb.platon.web.api.common.InputSanitizer;
 import de.vorb.platon.web.api.common.PoweredByResponseInterceptor;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.SerializationFeature;
+import org.owasp.html.HtmlPolicyBuilder;
+import org.owasp.html.PolicyFactory;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurerAdapter;
 
 @Configuration
+@ComponentScan(basePackageClasses = {PersistenceConfig.class, SecurityConfig.class})
 public class ApiConfig extends WebMvcConfigurerAdapter {
 
     @Bean
@@ -39,5 +47,21 @@ public class ApiConfig extends WebMvcConfigurerAdapter {
     @Override
     public void addInterceptors(InterceptorRegistry registry) {
         registry.addInterceptor(new PoweredByResponseInterceptor());
+    }
+
+    @Bean
+    public InputSanitizer htmlInputSanitizer(@Value("${platon.input.html_elements}") String allowedHtmlElements) {
+
+        final HtmlPolicyBuilder htmlPolicyBuilder = new HtmlPolicyBuilder()
+                .allowUrlProtocols("http", "https", "mailto")
+                .allowAttributes("href").onElements("a")
+                .allowAttributes("src", "width", "height", "alt").onElements("img")
+                .allowAttributes("class").onElements("div", "span");
+
+        htmlPolicyBuilder.allowElements(allowedHtmlElements.trim().split("\\s*,\\s*"));
+
+        final PolicyFactory htmlContentPolicy = htmlPolicyBuilder.toFactory();
+
+        return htmlContentPolicy::sanitize;
     }
 }
