+++ src/main/java/de/vorb/platon/web/api/controllers/CommentController.java
@@ -22,9 +22,9 @@ import de.vorb.platon.model.CommentStatus;
 import de.vorb.platon.persistence.CommentRepository;
 import de.vorb.platon.persistence.ThreadRepository;
 import de.vorb.platon.security.RequestVerifier;
-import de.vorb.platon.util.CommentConverter;
-import de.vorb.platon.util.CommentFilters;
-import de.vorb.platon.util.InputSanitizer;
+import de.vorb.platon.web.api.common.CommentConverter;
+import de.vorb.platon.web.api.common.CommentFilters;
+import de.vorb.platon.web.api.common.InputSanitizer;
 import de.vorb.platon.web.api.errors.RequestException;
 import de.vorb.platon.web.api.json.CommentJson;
 import de.vorb.platon.web.api.json.CommentListResultJson;
